Moved jpa entities to sharedobjects
