+++ ListSharedObjects/build.gradle
@@ -10,7 +10,15 @@ repositories {
 }
 
 dependencies {
+//    implementation 'jakarta.persistence:jakarta.persistence-api:3.1.0'
+//    implementation 'com.fasterxml.jackson.core:jackson-annotations:2.15.3'
+
+    compileOnly 'com.fasterxml.jackson.core:jackson-annotations:2.15.3'
+    compileOnly 'jakarta.persistence:jakarta.persistence-api:3.1.0'
     compileOnly 'org.projectlombok:lombok:1.18.30'
+//    annotationProcessor "jakarta.persistence:jakarta.persistence-api:3.1.0"
+    annotationProcessor 'com.fasterxml.jackson.core:jackson-annotations:2.15.3'
+    annotationProcessor 'jakarta.persistence:jakarta.persistence-api:3.1.0'
     annotationProcessor 'org.projectlombok:lombok:1.18.30'
 }
