+++ ShoppingListManager/src/main/java/org/shoppinglistmanager/entity/ShoppingList.java
@@ -1,31 +1,31 @@
-package org.shoppinglistmanager.entity;
-
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.util.ArrayList;
-import java.util.List;
-
-@Builder
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Entity
-public class ShoppingList {
-    @Id
-    @GeneratedValue(strategy =  GenerationType.AUTO)
-    private Integer id;
-    private String title;
-    private String description;
-
-    @OneToMany(mappedBy = "shoppingList",
-            cascade = CascadeType.ALL,
-            orphanRemoval = true)
-    private final List<Item> listOfItems = new ArrayList<>(); // TODO: Need to be set to final, and then be checked
-
-    public ShoppingList(String title, String description) {
-        this.title = title;
-        this.description = description;
-    }
-}
+//package org.shoppinglistmanager.entity;
+//
+//
+//import jakarta.persistence.*;
+//import lombok.*;
+//
+//import java.util.ArrayList;
+//import java.util.List;
+//
+//@Builder
+//@Data
+//@AllArgsConstructor
+//@NoArgsConstructor
+//@Entity
+//public class ShoppingList {
+//    @Id
+//    @GeneratedValue(strategy =  GenerationType.AUTO)
+//    private Integer id;
+//    private String title;
+//    private String description;
+//
+//    @OneToMany(mappedBy = "shoppingList",
+//            cascade = CascadeType.ALL,
+//            orphanRemoval = true)
+//    private final List<Item> listOfItems = new ArrayList<>(); // TODO: Need to be set to final, and then be checked
+//
+//    public ShoppingList(String title, String description) {
+//        this.title = title;
+//        this.description = description;
+//    }
+//}
