+++ ShoppingListManager/src/main/java/org/shoppinglistmanager/repository/ShoppingListRepository.java
@@ -1,6 +1,6 @@
 package org.shoppinglistmanager.repository;
 
-import org.shoppinglistmanager.entity.ShoppingList;
+import org.listobjects.entity.ShoppingList;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
